backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.19
pandas>=2.3.0
sqlparse>=0.5.0
groq>=0.13.0
ollama>=0.4.0
sentence-transformers>=3.3.0
scikit-learn>=1.6.0
pydantic-settings>=2.7.0
aiofiles>=24.1.0
passlib>=1.7.4
bcrypt>=4.2.0
pyjwt>=2.9.0
email-validator>=2.2.0
openpyxl>=3.1.5
